[WIP] Registry Setup, End Ores, License Headers, Docs
I did a lot of basic setup while I was waiting for support.
This is complete with license notices via Quilt Licenser, JavaDocs everywhere, and a full registry system complete with batch registry.

LICENSE has been modified to replace the copyright symbol with an ASCII-friendly equivalent and add a dash to "NON-INFRINGEMENT"